src/main/java/com/century/traineetask/api/OrderController.java
package com.century.traineetask.api;

import com.century.traineetask.model.Order;
import com.century.traineetask.service.OrderService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("api/v1/order")
public class OrderController {

    private OrderService orderService;

    @Autowired
    public OrderController(OrderService orderService) {
        this.orderService = orderService;
    }

    @GetMapping
    public List<Order> getAllOrders() {
        return orderService.getAllOrders();
    }

    @GetMapping(path = "/{id}")
    public ResponseEntity<Order> getOrderById(@PathVariable("id") Long orderId) {
        Order order = orderService.getOrderById(orderId);

        return ResponseEntity.ok().body(order);
    }

    @PostMapping
    public void createOrder(@RequestBody Order order) {
        orderService.createOrder(order);
    }

    @DeleteMapping(path = "/{id}")
    public void deleteOrder(@PathVariable("id") Long orderId) {
        orderService.deleteOrder(orderId);
    }

    @PutMapping(path = "/{id}")
    public void updateOrder(@PathVariable("id") Long orderId, @RequestBody Order order) {
        orderService.updateOrder(orderId, order);
    }

}
